server/requirements.txt
flask==3.1.0
bleak==0.22.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
APScheduler==3.10.4
python-dotenv==1.0.1
